+++ workspace/ims-webapp/src/main/java/th/go/excise/ims/ta/persistence/repository/TaPaperSvPriPerUnRepository.java
@@ -1,13 +1,9 @@
 
 package th.go.excise.ims.ta.persistence.repository;
 
-import java.math.BigDecimal;
 import th.co.baiwa.buckwaframework.common.persistence.repository.CommonJpaCrudRepository;
 import th.go.excise.ims.ta.persistence.entity.TaPaperSvPriPerUn;
 
-public interface TaPaperSvPriPerUnRepository
-    extends CommonJpaCrudRepository<TaPaperSvPriPerUn, BigDecimal>
-{
-
+public interface TaPaperSvPriPerUnRepository extends CommonJpaCrudRepository<TaPaperSvPriPerUn, Long> {
 
 }
