+++ workspace/ims-webapp/src/main/java/th/go/excise/ims/ta/persistence/repository/TaPaperSvTaxAmtAddRepository.java
@@ -1,13 +1,9 @@
 
 package th.go.excise.ims.ta.persistence.repository;
 
-import java.math.BigDecimal;
 import th.co.baiwa.buckwaframework.common.persistence.repository.CommonJpaCrudRepository;
 import th.go.excise.ims.ta.persistence.entity.TaPaperSvTaxAmtAdd;
 
-public interface TaPaperSvTaxAmtAddRepository
-    extends CommonJpaCrudRepository<TaPaperSvTaxAmtAdd, BigDecimal>
-{
-
+public interface TaPaperSvTaxAmtAddRepository extends CommonJpaCrudRepository<TaPaperSvTaxAmtAdd, Long> {
 
 }
